requirements/dev_requirements.txt
pytest
pylint
mindspore-dev
torch
numpy
requests
tqdm
ddt
transformers==2.0.0